infra: - split functions into separate files - fix specs
