+++ src/is-custom-import.ts
@@ -1,0 +1,5 @@
+import commander from 'commander';
+
+export function isCustomImport(literal: string): boolean {
+  return commander.userLibPrefixes.some((prefix) => literal.startsWith(prefix));
+}
