+++ src/merge-import-statements.ts
@@ -1,0 +1,4 @@
+export function mergeImportStatements(importStatementOne, importStatementTwo): string {
+  const importedValues = importStatementTwo.match('{(.*)}')[1];
+  return importStatementOne.replace('}', `,${importedValues}}`);
+}
