+++ src/types.ts
@@ -1,0 +1,6 @@
+export interface ImportCategories {
+  thirdPartyImportPot: Map<string, string>;
+  userLibraryPot: Map<string, string>;
+  differentUserModulePot: Map<string, string>;
+  sameModulePot: Map<string, string>;
+}
